Also replace squares

+++ node/server.js
@@ -18,25 +18,12 @@ client.on('message', async msg => {
   if (msg.author.bot) { return }
   if (CHANNEL_IDS.includes(msg.channel.id)) {
     const body = msg.content;
-    console.log(body);
-    if(body.includes('💚')) {
+    if(body.includes('💚') && body.includes('rdle')) {
       const newBody = body.replaceAll('💚', '💙');
+      const newNewBody = newBody.replaceAll('🟩', '🟦');
       const user = (await msg.guild.members.fetch(msg.author.id)).user;
-      await msg.channel.send(`${user}'s -rdle score:\n${newBody}`);
+      await msg.channel.send(`${user}'s -rdle score:\n${newNewBody}`);
       msg.delete();
     }
   }
-
-  // const commandBody = msg.content.substring(commandPrefix.length)
-  // const command = commandBody.split(' ')[0];
-  // const commandInput = commandBody.substring(command.length + 1);
-
-  // if (Object.keys(commands).includes(command.toLowerCase())) {
-  //   checkVoteLife();
-  //   commands[command.toLowerCase()](msg, commandInput);
-  // }
-  // else {
-  //   msg.channel.send(`Invalid Command: ${command}`);
-  // }
-
 });
